pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sid-training-free-reco"
version = "0.1.0"
description = "SID-based training-free LLM recommender system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "faiss-cpu>=1.13.2",
  "mlx-embeddings>=0.1.0",
  "mlx-lm>=0.31.1",
  "numpy>=2.3.0",
  "pandas>=2.3.0",
  "pydantic>=2.11.0",
  "python-dotenv>=1.2.2",
  "pyyaml>=6.0.2",
  "rich>=14.0.0",
  "typer>=0.16.0",
]

[project.scripts]
sid-reco = "sid_reco.cli:app"

[dependency-groups]
dev = [
  "mypy>=1.16.0",
  "pytest>=8.4.0",
  "pytest-cov>=7.0.0",
  "ruff>=0.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/sid_reco"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["sid_reco"]

[tool.uv]
package = true
cache-dir = ".uv_cache"
